port long poll test
