+++ clients/src/test/java/org/apache/kafka/clients/consumer/internals/AsyncKafkaConsumerTest.java
@@ -384,38 +384,6 @@ public class AsyncKafkaConsumerTest {
         }
     }
 
-    @Test
-    public void testLongPollWaitIsLimited() {
-        String topicName = "topic1";
-        consumer.subscribe(singletonList(topicName));
-
-        assertEquals(singleton(topicName), consumer.subscription());
-        assertTrue(consumer.assignment().isEmpty());
-
-        final int partition = 3;
-        final TopicPartition tp = new TopicPartition(topicName, partition);
-        final List<ConsumerRecord<String, String>> records = asList(
-            new ConsumerRecord<>(topicName, partition, 2, "key1", "value1"),
-            new ConsumerRecord<>(topicName, partition, 3, "key2", "value2")
-        );
-
-        // On the first iteration, return no data; on the second, return two records
-        doAnswer(invocation -> {
-            // Mock the subscription being assigned as the first fetch is collected
-            subscriptions.assignFromSubscribed(Collections.singleton(tp));
-            return Fetch.empty();
-        }).doAnswer(invocation -> {
-            return Fetch.forPartition(tp, records, true);
-        }).when(fetchCollector).collectFetch(any(FetchBuffer.class));
-
-        // And then poll for up to 10000ms, which should return 2 records without timing out
-        ConsumerRecords<?, ?> returnedRecords = consumer.poll(Duration.ofMillis(10000));
-        assertEquals(2, returnedRecords.count());
-
-        assertEquals(singleton(topicName), consumer.subscription());
-        assertEquals(singleton(tp), consumer.assignment());
-    }
-
     private void assertNoPendingWakeup(final WakeupTrigger wakeupTrigger) {
         assertNull(wakeupTrigger.getPendingTask());
     }
